Add remove tests endpoint

+++ src/main/java/com/codeless/api/automation/repository/ExecutionRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExecutionRepository extends JpaRepository<Execution, Integer> {}
+public interface ExecutionRepository extends JpaRepository<Execution, Long> {}
