+++ src/main/java/com/codeless/api/automation/repository/TestRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TestRepository extends JpaRepository<Test, Integer> {}
+public interface TestRepository extends JpaRepository<Test, Long> {}
